+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+### Fixed
+
+-   Fix the disable logging logic in `notecard_is_present` to not assert if `i2c_nrfx_twim` is being used instead of `i2c_nrfx_twi`.
+
 ## [1.4.2] - 2023-11-30
 
 ### Changed